remove original fluxstep and use the common flowmatchingpipeline
