+++ src/megatron/bridge/diffusion/models/flux/flux_step.py
@@ -12,10 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+"""
+FLUX Forward Step.
+
+This is a prototype showing how to integrate the FlowMatchingPipeline
+into Megatron's training flow, reusing the well-tested flow matching logic.
+"""
 
 import logging
-import math
-from functools import lru_cache, partial
+from functools import partial
 from typing import Iterable
 
 import torch
@@ -23,12 +28,18 @@ from megatron.core import parallel_state
 from megatron.core.models.common.vision_module.vision_module import VisionModule
 from megatron.core.utils import get_model_config
 
-from megatron.bridge.diffusion.models.flux.flow_matching.flux_inference_pipeline import FlowMatchEulerDiscreteScheduler
+from megatron.bridge.diffusion.common.flow_matching.flow_matching_pipeline import FlowMatchingPipeline
+from megatron.bridge.diffusion.models.flux.flow_matching.flux_adapter import MegatronFluxAdapter
 from megatron.bridge.training.losses import masked_next_token_loss
 from megatron.bridge.training.state import GlobalState
 
 
 logger = logging.getLogger(__name__)
+
+
+# =============================================================================
+# Megatron Forward Step
+# =============================================================================
 
 
 def flux_data_step(dataloader_iter, store_in_state=False):
@@ -66,21 +77,18 @@ def flux_data_step(dataloader_iter, store_in_state=False):
 
 
 class FluxForwardStep:
-    """Forward step for FLUX diffusion model training.
+    """
+    Forward step for FLUX using FlowMatchingPipeline.
 
-    This class handles the forward pass during training, including:
-    - Timestep sampling using flow matching
-    - Noise injection with latent packing
-    - Model prediction
-    - Loss computation
-
+    This class demonstrates how to integrate the FlowMatchingPipeline
     Args:
         timestep_sampling: Method for sampling timesteps ("logit_normal", "uniform", "mode").
         logit_mean: Mean for logit-normal sampling.
         logit_std: Standard deviation for logit-normal sampling.
-        mode_scale: Scale for mode sampling.
+        flow_shift: Shift parameter for timestep transformation (default: 1.0 for FLUX).
         scheduler_steps: Number of scheduler training steps.
         guidance_scale: Guidance scale for FLUX-dev models.
+        use_loss_weighting: Whether to apply flow-based loss weighting.
     """
 
     def __init__(
@@ -88,24 +96,44 @@ class FluxForwardStep:
         timestep_sampling: str = "logit_normal",
         logit_mean: float = 0.0,
         logit_std: float = 1.0,
-        mode_scale: float = 1.29,
+        flow_shift: float = 1.0,  # FLUX uses shift=1.0 typically
         scheduler_steps: int = 1000,
         guidance_scale: float = 3.5,
+        use_loss_weighting: bool = False,  # FLUX typically doesn't use loss weighting
     ):
-        self.timestep_sampling = timestep_sampling
-        self.logit_mean = logit_mean
-        self.logit_std = logit_std
-        self.mode_scale = mode_scale
-        self.scheduler_steps = scheduler_steps
-        self.guidance_scale = guidance_scale
         self.autocast_dtype = torch.bfloat16
-        # Initialize scheduler for timestep/sigma computations
-        self.scheduler = FlowMatchEulerDiscreteScheduler(num_train_timesteps=scheduler_steps)
+
+        # Create the FlowMatchingPipeline with Megatron adapter
+        adapter = MegatronFluxAdapter(guidance_scale=guidance_scale)
+
+        self.pipeline = FlowMatchingPipeline(
+            model_adapter=adapter,
+            num_train_timesteps=scheduler_steps,
+            timestep_sampling=timestep_sampling,
+            flow_shift=flow_shift,
+            logit_mean=logit_mean,
+            logit_std=logit_std,
+            sigma_min=0.0,
+            sigma_max=1.0,
+            use_loss_weighting=use_loss_weighting,
+            cfg_dropout_prob=0.0,  # No CFG dropout in Megatron training
+            log_interval=100,
+            summary_log_interval=10,
+        )
+
+        logger.info(
+            f"FluxForwardStep initialized with:\n"
+            f"  - Timestep sampling: {timestep_sampling}\n"
+            f"  - Flow shift: {flow_shift}\n"
+            f"  - Guidance scale: {guidance_scale}\n"
+            f"  - Loss weighting: {use_loss_weighting}"
+        )
 
     def __call__(
         self, state: GlobalState, data_iterator: Iterable, model: VisionModule
     ) -> tuple[torch.Tensor, partial]:
-        """Forward training step.
+        """
+        Forward training step using FlowMatchingPipeline.
 
         Args:
             state: Global state for the run.
@@ -118,7 +146,7 @@ class FluxForwardStep:
         timers = state.timers
         straggler_timer = state.straggler_timer
 
-        config = get_model_config(model)
+        config = get_model_config(model)  # noqa: F841
 
         timers("batch-generator", log_level=2).start()
 
@@ -132,289 +160,111 @@ class FluxForwardStep:
         check_for_nan_in_loss = state.cfg.rerun_state_machine.check_for_nan_in_loss
         check_for_spiky_loss = state.cfg.rerun_state_machine.check_for_spiky_loss
 
-        # Run diffusion training step
+        # Prepare batch for FlowMatchingPipeline
+        # Map Megatron keys to FlowMatchingPipeline expected keys
+        pipeline_batch = self._prepare_batch_for_pipeline(batch)
+
+        # Run the pipeline step
         with straggler_timer:
             if parallel_state.is_pipeline_last_stage():
-                output_tensor, loss, loss_mask = self._training_step(model, batch, config)
+                output_tensor, loss, loss_mask = self._training_step_with_pipeline(model, pipeline_batch)
+                # loss_mask is already created correctly in _training_step_with_pipeline
                 batch["loss_mask"] = loss_mask
             else:
-                output_tensor = self._training_step(model, batch, config)
+                # For non-final pipeline stages, we still need to run the model
+                # but loss computation happens only on the last stage
+                output_tensor = self._training_step_with_pipeline(model, pipeline_batch)
+                loss_mask = None
 
-        loss = output_tensor
-        if "loss_mask" not in batch or batch["loss_mask"] is None:
-            loss_mask = torch.ones_like(loss)
-        else:
-            loss_mask = batch["loss_mask"]
+        # Use the loss_mask from training step (already has correct shape)
+        if loss_mask is None:
+            # This should only happen for non-final pipeline stages
+            loss_mask = torch.ones(1, device="cuda")
 
         loss_function = self._create_loss_function(loss_mask, check_for_nan_in_loss, check_for_spiky_loss)
 
         return output_tensor, loss_function
 
-    def _training_step(
-        self, model: VisionModule, batch: dict, config
+    def _prepare_batch_for_pipeline(self, batch: dict) -> dict:
+        """
+        Prepare Megatron batch for FlowMatchingPipeline.
+
+        Maps Megatron batch keys to FlowMatchingPipeline expected format:
+        - latents -> image_latents (for consistency)
+        - Keeps prompt_embeds, pooled_prompt_embeds, text_ids as-is
+        """
+        pipeline_batch = {
+            "image_latents": batch["latents"],  # Map to FlowMatchingPipeline expected key
+            "prompt_embeds": batch.get("prompt_embeds"),
+            "pooled_prompt_embeds": batch.get("pooled_prompt_embeds"),
+            "text_ids": batch.get("text_ids"),
+            "data_type": "image",  # FLUX is for image generation
+        }
+
+        # Copy any additional keys
+        for key in batch:
+            if key not in pipeline_batch and key != "latents":
+                pipeline_batch[key] = batch[key]
+
+        return pipeline_batch
+
+    def _training_step_with_pipeline(
+        self, model: VisionModule, batch: dict
     ) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor] | torch.Tensor:
-        """Perform single training step with flow matching.
+        """
+        Perform single training step using FlowMatchingPipeline.
 
         Args:
             model: The FLUX model.
-            batch: Data batch containing latents and text embeddings.
-            config: Model configuration.
+            batch: Data batch prepared for pipeline.
 
         Returns:
             On last pipeline stage: tuple of (output_tensor, loss, loss_mask).
-            On other stages: hidden_states tensor.
+            On other stages: output tensor.
         """
-        # Get latents from batch - expected in [B, C, H, W] format
-        if "latents" in batch:
-            latents = batch["latents"]
-        else:
-            raise ValueError("Expected 'latents' in batch. VAE encoding should be done in data preprocessing.")
+        device = torch.device("cuda")
+        dtype = self.autocast_dtype
 
-        # Prepare image latents with flow matching noise
-        (
-            latents,
-            noise,
-            packed_noisy_model_input,
-            latent_image_ids,
-            guidance_vec,
-            timesteps,
-        ) = self.prepare_image_latent(latents, model)
+        # Pass model in batch so adapter can check for guidance support
+        batch["_model"] = model
 
-        # Get text embeddings (precached)
-        if "prompt_embeds" in batch:
-            prompt_embeds = batch["prompt_embeds"].transpose(0, 1)
-            pooled_prompt_embeds = batch["pooled_prompt_embeds"]
-            text_ids = batch["text_ids"]
-        else:
-            raise ValueError("Expected precached text embeddings in batch.")
-
-        # Forward pass
-        with torch.amp.autocast(
-            "cuda", enabled=self.autocast_dtype in (torch.half, torch.bfloat16), dtype=self.autocast_dtype
-        ):
-            noise_pred = model(
-                img=packed_noisy_model_input,
-                txt=prompt_embeds,
-                y=pooled_prompt_embeds,
-                timesteps=timesteps / 1000,
-                img_ids=latent_image_ids,
-                txt_ids=text_ids,
-                guidance=guidance_vec,
+        with torch.amp.autocast("cuda", enabled=dtype in (torch.half, torch.bfloat16), dtype=dtype):
+            # Run the FlowMatchingPipeline step (global_step defaults to 0)
+            weighted_loss, average_weighted_loss, loss_mask, metrics = self.pipeline.step(
+                model=model,
+                batch=batch,
+                device=device,
+                dtype=dtype,
             )
 
-            # Unpack predictions for loss computation
-            noise_pred = self._unpack_latents(
-                noise_pred.transpose(0, 1),
-                latents.shape[2],
-                latents.shape[3],
-            ).transpose(0, 1)
+        # Clean up temporary model reference
+        batch.pop("_model", None)
 
-            # Flow matching target: v = noise - latents (velocity formulation)
-            target = noise - latents
+        if parallel_state.is_pipeline_last_stage():
+            # Match original implementation's reduction pattern
+            # Original does: loss = mse(..., reduction="none"), then output_tensor = mean(loss, dim=-1)
+            # This keeps most dimensions and only reduces the last one
+            # But FlowMatchingPipeline returns full loss, so we reduce to match expected shape
 
-            # MSE loss
-            loss = torch.nn.functional.mse_loss(noise_pred.float(), target.float(), reduction="none")
-            output_tensor = torch.mean(loss, dim=-1)
+            # For FLUX with images: weighted_loss is [B, C, H, W]
+            # Original pattern: mean over spatial dimensions -> [B, C] or similar
+            # But Megatron expects a 1D tensor per sample, so reduce to [B]
+            output_tensor = torch.mean(weighted_loss, dim=list(range(1, weighted_loss.ndim)))
 
-            # Create loss mask (all ones for now)
+            # Always create a fresh loss_mask matching output_tensor shape
+            # Ignore any loss_mask from batch as it may have incompatible shape
             loss_mask = torch.ones_like(output_tensor)
 
-            return output_tensor, loss, loss_mask
-
-        # else:
-        #     hidden_states = model(
-        #         img=packed_noisy_model_input,
-        #         txt=prompt_embeds,
-        #         y=pooled_prompt_embeds,
-        #         timesteps=timesteps / 1000,
-        #         img_ids=latent_image_ids,
-        #         txt_ids=text_ids,
-        #         guidance=guidance_vec,
-        #     )
-        #     return hidden_states
-
-    def prepare_image_latent(self, latents: torch.Tensor, model: VisionModule):
-        """Prepare image latents with flow matching noise.
-
-        Args:
-            latents: Input latent tensor [B, C, H, W].
-            model: The FLUX model (for guidance_embed config).
-
-        Returns:
-            Tuple of (latents, noise, packed_noisy_input, latent_image_ids, guidance, timesteps).
-        """
-        latent_image_ids = self._prepare_latent_image_ids(
-            latents.shape[0],
-            latents.shape[2],
-            latents.shape[3],
-            latents.device,
-            latents.dtype,
-        )
-
-        noise = torch.randn_like(latents, device=latents.device, dtype=latents.dtype)
-        batch_size = latents.shape[0]
-        u = self.compute_density_for_timestep_sampling(
-            self.timestep_sampling,
-            batch_size,
-        )
-        indices = (u * self.scheduler.num_train_timesteps).long()
-        timesteps = self.scheduler.timesteps[indices].to(device=latents.device)
-
-        sigmas = self.scheduler.sigmas.to(device=latents.device, dtype=latents.dtype)
-        scheduler_timesteps = self.scheduler.timesteps.to(device=latents.device)
-        step_indices = [(scheduler_timesteps == t).nonzero().item() for t in timesteps]
-        timesteps = timesteps.to(dtype=latents.dtype)
-        sigma = sigmas[step_indices].flatten()
-
-        while len(sigma.shape) < latents.ndim:
-            sigma = sigma.unsqueeze(-1)
-
-        noisy_model_input = (1.0 - sigma) * latents + sigma * noise
-        packed_noisy_model_input = self._pack_latents(
-            noisy_model_input,
-            batch_size=latents.shape[0],
-            num_channels_latents=latents.shape[1],
-            height=latents.shape[2],
-            width=latents.shape[3],
-        )
-
-        # Guidance embedding (for FLUX-dev)
-        if hasattr(model, "guidance_embed") and model.guidance_embed:
-            guidance_vec = torch.full(
-                (noisy_model_input.shape[0],),
-                self.guidance_scale,
-                device=latents.device,
-                dtype=latents.dtype,
-            )
+            return output_tensor, average_weighted_loss, loss_mask
         else:
-            guidance_vec = None
-
-        return (
-            latents.transpose(0, 1),
-            noise.transpose(0, 1),
-            packed_noisy_model_input.transpose(0, 1),
-            latent_image_ids,
-            guidance_vec,
-            timesteps,
-        )
-
-    def compute_density_for_timestep_sampling(
-        self,
-        weighting_scheme: str,
-        batch_size: int,
-        logit_mean: float = None,
-        logit_std: float = None,
-        mode_scale: float = None,
-    ) -> torch.Tensor:
-        """Compute the density for sampling the timesteps when doing SD3 training.
-
-        Courtesy: This was contributed by Rafie Walker in https://github.com/huggingface/diffusers/pull/8528.
-        SD3 paper reference: https://arxiv.org/abs/2403.03206v1.
-
-        Args:
-            weighting_scheme: Sampling scheme ("logit_normal", "mode", or "uniform").
-            batch_size: Number of samples in batch.
-            logit_mean: Mean for logit-normal sampling.
-            logit_std: Standard deviation for logit-normal sampling.
-            mode_scale: Scale for mode sampling.
-
-        Returns:
-            Tensor of sampled u values in [0, 1].
-        """
-        # Use instance defaults if not provided
-        logit_mean = logit_mean if logit_mean is not None else self.logit_mean
-        logit_std = logit_std if logit_std is not None else self.logit_std
-        mode_scale = mode_scale if mode_scale is not None else self.mode_scale
-
-        if weighting_scheme == "logit_normal":
-            # See 3.1 in the SD3 paper ($rf/lognorm(0.00,1.00)$)
-            u = torch.normal(mean=logit_mean, std=logit_std, size=(batch_size,), device="cpu")
-            u = torch.nn.functional.sigmoid(u)
-        elif weighting_scheme == "mode":
-            u = torch.rand(size=(batch_size,), device="cpu")
-            u = 1 - u - mode_scale * (torch.cos(math.pi * u / 2) ** 2 - 1 + u)
-        else:
-            u = torch.rand(size=(batch_size,), device="cpu")
-        return u
-
-    @lru_cache
-    def _prepare_latent_image_ids(
-        self, batch_size: int, height: int, width: int, device: torch.device, dtype: torch.dtype
-    ) -> torch.Tensor:
-        """Prepare latent image IDs for positional encoding.
-
-        Args:
-            batch_size: Number of samples.
-            height: Latent height.
-            width: Latent width.
-            device: Target device.
-            dtype: Target dtype.
-
-        Returns:
-            Tensor of shape [B, (H/2)*(W/2), 3] with position IDs.
-        """
-        latent_image_ids = torch.zeros(height // 2, width // 2, 3)
-        latent_image_ids[..., 1] = latent_image_ids[..., 1] + torch.arange(height // 2)[:, None]
-        latent_image_ids[..., 2] = latent_image_ids[..., 2] + torch.arange(width // 2)[None, :]
-
-        latent_image_id_height, latent_image_id_width, latent_image_id_channels = latent_image_ids.shape
-
-        latent_image_ids = latent_image_ids[None, :].repeat(batch_size, 1, 1, 1)
-        latent_image_ids = latent_image_ids.reshape(
-            batch_size, latent_image_id_height * latent_image_id_width, latent_image_id_channels
-        )
-
-        return latent_image_ids.to(device=device, dtype=dtype, non_blocking=True)
-
-    def _pack_latents(
-        self, latents: torch.Tensor, batch_size: int, num_channels_latents: int, height: int, width: int
-    ) -> torch.Tensor:
-        """Pack latents for FLUX processing.
-
-        Rearranges [B, C, H, W] -> [B, (H/2)*(W/2), C*4].
-
-        Args:
-            latents: Input tensor [B, C, H, W].
-            batch_size: Batch size.
-            num_channels_latents: Number of latent channels.
-            height: Latent height.
-            width: Latent width.
-
-        Returns:
-            Packed tensor [B, num_patches, C*4].
-        """
-        latents = latents.view(batch_size, num_channels_latents, height // 2, 2, width // 2, 2)
-        latents = latents.permute(0, 2, 4, 1, 3, 5)
-        latents = latents.reshape(batch_size, (height // 2) * (width // 2), num_channels_latents * 4)
-        return latents
-
-    def _unpack_latents(self, latents: torch.Tensor, height: int, width: int) -> torch.Tensor:
-        """Unpack latents from FLUX format.
-
-        Rearranges [B, num_patches, C*4] -> [B, C, H, W].
-
-        Args:
-            latents: Packed tensor [B, num_patches, C*4].
-            height: Target height.
-            width: Target width.
-
-        Returns:
-            Unpacked tensor [B, C, H, W].
-        """
-        batch_size, num_patches, channels = latents.shape
-
-        # Adjust h and w for patching
-        latents = latents.view(batch_size, height // 2, width // 2, channels // 4, 2, 2)
-        latents = latents.permute(0, 3, 1, 4, 2, 5)
-
-        latents = latents.reshape(batch_size, channels // 4, height, width)
-
-        return latents
+            # For intermediate stages, return the tensor for pipeline communication
+            return weighted_loss
 
     def _create_loss_function(
         self, loss_mask: torch.Tensor, check_for_nan_in_loss: bool, check_for_spiky_loss: bool
     ) -> partial:
-        """Create a partial loss function with the specified configuration.
+        """
+        Create a partial loss function with the specified configuration.
 
         Args:
             loss_mask: Used to mask out some portions of the loss.
